Update: Modify product api

+++ src/main/java/com/yas/system/catalog/internal/dto/request/CategoryRequest.java
@@ -2,6 +2,6 @@ package com.yas.system.catalog.internal.dto.request;
 
 public record CategoryRequest(
         String name,
-        Long parentId
+        Integer parentId
 ) {
 }
