+++ src/main/java/com/yas/system/catalog/internal/repository/CategoryRepository.java
@@ -10,16 +10,14 @@ import java.util.Optional;
 
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-
-
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 
     @Query("""
         select c 
         from Category c 
         where c.name = :name and (c.id != :id or :id is null)
         """)
-    Optional<Category> checkExited(String name, Long id);
+    Optional<Category> checkExited(String name, Integer id);
 
 
     @Query("""
@@ -35,5 +33,5 @@ public interface CategoryRepository extends JpaRepository<Category, Long> {
         left join fetch c.children
         where c.id = :id 
     """)
-    Optional<Category> findByIdCustom(Long id);
+    Optional<Category> findByIdCustom(Integer id);
 }
